generics 0

+++ Generics_Exercises/armazem/Comida.java
@@ -1,10 +1,10 @@
 
 public class Comida {
     private String nome;
-    private int calorias;
+    private double calorias;
     private double preco;
     
-    public Comida(String nome, int calorias, double preco) {
+    public Comida(String nome, double calorias, double preco) {
         this.nome = nome;
         this.calorias = calorias;
         this.preco = preco;
@@ -12,7 +12,7 @@ public class Comida {
 
    @Override
    public String toString(){
-    return String.format("[%s] %d R$ %.6f", nome, calorias, preco);
+    return String.format("[%s] %.6f R$ %.6f", nome, calorias, preco);
    } 
 
 }
